+++ uuid/uuid_test.go
@@ -28,11 +28,19 @@ var _ = Suite(&UUIDSuite{})
 func (s *UUIDSuite) TestUUID4(c *C) {
 	c.Assert(UUID4(), HasLen, 16)
 	c.Assert(UUID4().String(), Not(Equals), "00000000-0000-0000-0000-000000000000")
+	c.Assert(UUID4().IsZero(), Equals, false)
 }
 
 func (s *UUIDSuite) TestUUID5(c *C) {
 	c.Assert(UUID5(NsURL, "TEST"), HasLen, 16)
 	c.Assert(UUID5(NsURL, "TEST").String(), Not(Equals), "00000000-0000-0000-0000-000000000000")
+	c.Assert(UUID5(NsURL, "TEST").IsZero(), Equals, false)
+}
+
+func (s *UUIDSuite) TestUUID7(c *C) {
+	c.Assert(UUID7(), HasLen, 16)
+	c.Assert(UUID7().String(), Not(Equals), "00000000-0000-0000-0000-000000000000")
+	c.Assert(UUID7().IsZero(), Equals, false)
 }
 
 func (s *UUIDSuite) BenchmarkUUID4(c *C) {