+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+### [12.127.0](https://kaos.sh/ek/12.127.0)
+
+- `[options]` Added method `Delete`
+- `[usage]` Added support for optional arguments to usage info
+- `[uuid]` Added [UUID7](https://uuid7.com) generator
+- `[options]` Code refactoring
+
 ### [12.126.1](https://kaos.sh/ek/12.126.1)
 
 - `[usage]` Added `Info.WrapLen` option for text wrapping configuration